pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Melnik <melnik721ivan@gmail.com>"]
readme = "README.md"

packages = [
  { include = "gendiff" },
]


[tool.poetry.dependencies]
python = "^3.10"
argparse = "^1.4.0"
pyyaml = "^6.0.1"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.generate_diff:main"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pytest = "^8.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
